requirements.txt
django==5.2.9
drf-spectacular[sidecar]==0.28.0
django-redis==6.0.0
django-db-connection-pool==1.2.6
django-mptt==0.17.0
psycopg[binary]==3.2.9
python-dotenv==1.1.1
uuid-utils==0.11.0
captcha==0.7.1
pytz==2025.2
psutil==7.0.0
cffi==1.17.1
beautifulsoup4==4.13.4
jieba==0.42.1
langchain-openai==0.3.35
langchain-anthropic==0.3.22
langchain-community==0.3.31
langchain-deepseek==0.1.4
langchain-google-genai==2.1.12
langchain-mcp-adapters==0.1.13
langchain-huggingface==0.3.1
langchain-ollama==0.3.10
langchain-core==0.3.81
langchain-aws==0.2.28
langgraph==0.5.3
torch==2.8.0
sentence-transformers==5.0.0
qianfan==0.4.12.3
zhipuai==2.1.5.20250708
volcengine-python-sdk[ark]==4.0.5
boto3==1.39.4
tencentcloud-sdk-python==3.0.1420
xinference-client==1.7.1.post1
anthropic==0.74.1
dashscope==1.23.8
celery[sqlalchemy]==5.5.3
django-celery-beat==2.8.1
celery-once==3.0.1
django-apscheduler==0.7.0
html2text==2025.4.15
openpyxl==3.1.5
python-docx==1.2.0
xlrd==2.0.2
xlwt==1.3.0
pymupdf==1.26.3
pypdf==6.4.0
pydub==0.25.1
pysilk==0.0.1
gunicorn==23.0.0
python-daemon==3.1.2
websockets==15.0.1
pylint==3.3.7
cohere==5.17.0
jsonpath-ng==1.7.0
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-pgvector==0.4.0
